fix(intfsorch): resolve -Werror=conversion for stoul to uint32_t
Cast std::stoul() result to uint32_t explicitly to satisfy
-Werror=conversion. Values are bounded by VLAN range (1-4094) so
no actual truncation risk.

+++ orchagent/intfsorch.cpp
@@ -116,7 +116,7 @@ IntfsOrch::IntfsOrch(DBConnector *db, string tableName, VRFOrch *vrf_orch, DBCon
     {
         try
         {
-            m_maxSviCapacity = std::stoul(maxSviStr);
+            m_maxSviCapacity = static_cast<uint32_t>(std::stoul(maxSviStr));
             SWSS_LOG_NOTICE("SVI/RIF creation hard-cap set to %u (from CRM|Config.max_svi_capacity)", m_maxSviCapacity);
         }
         catch (...)
